Trace method calls (#15)
* #1: Trace method calls.

Not working for some methods.

* #1: Fix exception when transforming methods with one parameter

* #1: Cleanup and comments

* #1: Start DynamicClassDefiner

* #1: Implement DynamicClassDefiner

* #1: Add a simple trace processor thread

* #1: Make the trace processor a daemon thread

* #1: Spotless

* #1: Fix DynamicClass visibility

* #1: Pull java bin dir from the environment

* #1: Convert DynamicClassDefinerTest to kotlin

* #1: Add detekt

* #1: Apply GPL license

* #1: Add method location to the generated container classes

* WIP

* #1: Implement serialization for method traces

* #1: Fix usage of DynamicClassDefiner. Start working on UI.

There is some bug relating to the remaining length part of the trace being too small sometimes.

* #1: Replace remainingLength with number of arguments in the method traces to fix trace parsing.

* #1: Implement parsing for all types

* #1: Display traces as a list

* #1: Cleanup

* #1: Fix DynamicClassDefinerTest.kt

* #1: Documentation

* #1: Cleanup index parsing

* #1: Refactor packages

* Adding kotest

* Move to kts build file and fix kotest dependencies

* #1: More refactoring and implement parseNoArgMethodCall

* #1: Finish testing TraceIterator

* #1: Store string values in traces

* #1: Trace line numbers

* #1: Cleanup

* #1: Fix tracing of class names correlated with line numbers

* #1: Working on proper integration testing

* #1: Establish proper integration testing and convert two of the TraceIterator tests.

* #1: Move version numbers from gradle.properties to Versions.kt.

* #1: Convert remaining TraceIterator tests to the new format

* #1: Remove unused code

* #1: Remove old method tracing section from the readme

* #1: Extract the agent lib into a temp dir when running tests

* #1: Debug Windows failure

* #1: Debug Windows 2

* Revert "#1: Debug Windows 2"

This reverts commit 51f005d8

* #1: Fix Jar extraction logic so it doesn't extract the Jar each call.

+++ agent/src/main/java/com/octogonapus/omj/agent/AgentLibJarExtractor.java
@@ -1,12 +1,30 @@
+/*
+ * This file is part of OMJ.
+ *
+ * OMJ is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * OMJ is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with OMJ.  If not, see <https://www.gnu.org/licenses/>.
+ */
 package com.octogonapus.omj.agent;
 
+import com.octogonapus.omj.util.Util;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
-import java.nio.file.Paths;
 
-public class AgentLibJarExtractor {
+public final class AgentLibJarExtractor {
+
+  private static File agentLibAllJar = null;
 
   /**
    * Extracts the agent-lib jar from our jar and into a file.
@@ -20,20 +38,17 @@ public class AgentLibJarExtractor {
         throw new IOException("Could not locate agent-lib-all resource.");
       }
 
-      final var file =
-          Paths.get(System.getProperty("user.home"), ".OMJ", "agent-lib-all.jar").toFile();
-
-      // The lib contents can change all the time so we should re-extract it each time
-      file.deleteOnExit();
-      if (file.exists() && !file.delete()) {
-        throw new IOException("Failed to delte file " + file.getPath());
-      } else if (!file.getParentFile().mkdirs() && !file.getParentFile().exists()) {
-        throw new IOException("Failed to make parent dirs for file " + file.getPath());
+      // Extract the Jar and save it to agentLibAllJar so that subsequent calls do not try to
+      // extract the Jar again.
+      if (agentLibAllJar == null) {
+        final var file =
+            Files.createTempFile(Util.getAgentLibJarDir(), "agent-lib-all_", ".jar").toFile();
+        file.deleteOnExit();
+        copyStreamToFile(file, agentLib);
+        agentLibAllJar = file;
       }
 
-      copyStreamToFile(file, agentLib);
-
-      return file;
+      return agentLibAllJar;
     }
   }
